refactor: move files around so they are easier to understand
